requirements.txt
discord.py>=2.3.0
yt-dlp>=2023.12.30
ffmpeg-python>=0.2.0
ffmpeg
asyncio
python-dotenv>=1.0.0
Flask>=3.0.0
Pillow>=10.0.0
aiohttp>=3.9.0
PyNaCl>=1.5.0
discord.py[voice]
async-timeout>=4.0.0
requests>=2.31.0
schedule>=1.2.0